requirements.txt
Flask~=3.1.0
flask-cors~=5.0.1
requests~=2.32.3
python-dotenv~=1.0.1
psycopg2~=2.9.10
pandas~=2.2.3
beautifulsoup4~=4.13.3